Refactor settings toolbar to use MUI styling framework.
Also splits the module into a single module per component.
